Renamed package and fixed import GoeCharger error
