+++ goecharger_api_lite/__init__.py
@@ -1,0 +1,1 @@
+from .goecharger_api_lite import GoeCharger  # noqa
