+++ goecharger_api_lite/goecharger_api_lite.py
@@ -8,7 +8,7 @@ from typing import Union, Tuple, Any, Dict, Optional, Coroutine
 import aiohttp
 from aiohttp import ClientTimeout, ClientError, ContentTypeError
 
-from goecharger.excpetion import GoeChargerError
+from goecharger_api_lite.excpetion import GoeChargerError
 
 
 class GoeCharger:
