+++ setup.py
@@ -1,4 +1,4 @@
-from setuptools import setup  # type: ignore
+from setuptools import setup # type: ignore
 
 # read the contents of README.md file
 from pathlib import Path
@@ -8,7 +8,7 @@ long_description = (this_directory / "README.md").read_text()
 setup(
     name='goecharger-api-lite',
     version='1.1.1',
-    packages=['goecharger'],
+    packages=['goecharger_api_lite'],
     url='https://github.com/bkogler/goecharger-api-lite',
     license='MIT',
     author='Bernhard Kogler',
